implements rate limiting and redis caching

+++ olp-courses/src/main/java/ward/peter/olp_courses/CourseRepository.java
@@ -2,4 +2,6 @@ package ward.peter.olp_courses;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseRepository extends JpaRepository<Course, Integer> {}
+import java.util.List;
+
+public interface CourseRepository extends JpaRepository<Course, Long> {}
